chore: I give up, writing this wg/overlay is driving me crazy
Does it even make sense to use this, instead of
ssh reverse port forwarding?

My vandrare PoC can even load balance clients over multiple
ssh reverse port forwarding. Besides, it is probably easier to
spin up a central wireguard guard server that could provide
UDP packet forwarding.

If one cannot open `a single` UDP port for wireguard traffic,
it makes no sense to expect high performance.

Also, the use-case here is to overlay an ssh session anyway,
and allow a TCP-like stream.

An alternative would be to use quick-go instead of SSH,
but then we land on the same issue of people not wanting
UDP...

The point is, I am tired and this is leading to nowhere.